PR #487 r12: share isLoopbackHostname between main + renderer
Codex P2 round 11: the renderer's localhost-detection regex in
PrintLabelDialog covered fewer shapes than the main-process validator
in electron/main.ts. Opening the web UI through `localhost.`, `[::]`,
or an IPv4-mapped IPv6 address would let `deepLinkAvailable` go true
and the URL-mode QR would still encode an unscannable local origin —
even though the equivalent string typed into Settings → Public URL
would have been rejected.

Fix is structural: extract the helper to a shared location so both
sides import the same code.

  electron/loopback-host.ts → src/lib/loopbackHost.ts (moved)
  + isLoopbackHostname (verbatim)
  + isLoopbackUrl(urlString) convenience wrapper for the renderer's
    `window.location.origin` check

electron/main.ts imports via relative path `../src/lib/loopbackHost`
(esbuild bundles it inline; the file is pure regex/string ops with
no Node OR browser deps). PrintLabelDialog imports via the existing
`@/lib/loopbackHost` alias.

The previous renderer regex is gone — replaced with
`isLoopbackUrl(origin)` so the UX gate and the security boundary
can't drift again. Tests moved to tests/loopbackHost.test.ts with
matching import path; 49 cases still pass.

Full suite: 1671 tests passing, lint + tsc clean, electron main
bundle still compiles.
